+++ bindings/matrix-sdk-crypto-js/src/verification.rs
@@ -7,6 +7,7 @@ use futures_util::StreamExt;
 #[cfg(feature = "qrcode")]
 use js_sys::Uint8ClampedArray;
 use js_sys::{Array, Function, JsString, Promise};
+use matrix_sdk_crypto::VerificationRequestState;
 use ruma::events::key::verification::{
     cancel::CancelCode as RumaCancelCode, VerificationMethod as RumaVerificationMethod,
 };
@@ -183,32 +184,32 @@ impl Sas {
     }
 
     /// Does this verification flow support displaying emoji for the
-    /// short authentication string.
+    /// short authentication string?
     #[wasm_bindgen(js_name = "supportsEmoji")]
     pub fn supports_emoji(&self) -> bool {
         self.inner.supports_emoji()
     }
 
-    /// Did this verification flow start from a verification request.
+    /// Did this verification flow start from a verification request?
     #[wasm_bindgen(js_name = "startedFromRequest")]
     pub fn started_from_request(&self) -> bool {
         self.inner.started_from_request()
     }
 
-    /// Is this a verification that is veryfying one of our own
-    /// devices.
+    /// Is this a verification that is verifying one of our own
+    /// devices?
     #[wasm_bindgen(js_name = "isSelfVerification")]
     pub fn is_self_verification(&self) -> bool {
         self.inner.is_self_verification()
     }
 
-    /// Have we confirmed that the short auth string matches.
+    /// Have we confirmed that the short auth string matches?
     #[wasm_bindgen(js_name = "haveWeConfirmed")]
     pub fn have_we_confirmed(&self) -> bool {
         self.inner.have_we_confirmed()
     }
 
-    /// Has the verification been accepted by both parties.
+    /// Has the verification been accepted by both parties?
     #[wasm_bindgen(js_name = "hasBeenAccepted")]
     pub fn has_been_accepted(&self) -> bool {
         self.inner.has_been_accepted()
@@ -221,7 +222,8 @@ impl Sas {
         self.inner.cancel_info().map(Into::into)
     }
 
-    /// Did we initiate the verification flow.
+    /// True if we initiated the verification flow (ie, we sent the
+    /// `m.key.verification.request`).
     #[wasm_bindgen(js_name = "weStarted")]
     pub fn we_started(&self) -> bool {
         self.inner.we_started()
@@ -229,9 +231,9 @@ impl Sas {
 
     /// Accept the SAS verification.
     ///
-    /// This does nothing if the verification was already accepted,
-    /// otherwise it returns an `AcceptEventContent` that needs to be
-    /// sent out.
+    /// This does nothing (and returns `undefined`) if the verification was
+    /// already accepted, otherwise it returns an `OutgoingRequest`
+    /// that needs to be sent out.
     pub fn accept(&self) -> Result<JsValue, JsError> {
         self.inner
             .accept()
@@ -275,6 +277,9 @@ impl Sas {
     }
 
     /// Cancel the verification.
+    ///
+    /// Returns either an `OutgoingRequest` which should be sent out, or
+    /// `undefined` if the verification is already cancelled.
     pub fn cancel(&self) -> Result<JsValue, JsError> {
         self.inner
             .cancel()
@@ -288,6 +293,9 @@ impl Sas {
     /// Cancel the verification.
     ///
     /// This cancels the verification with given code.
+    ///
+    /// Returns either an `OutgoingRequest` which should be sent out, or
+    /// `undefined` if the verification is already cancelled.
     #[wasm_bindgen(js_name = "cancelWithCode")]
     pub fn cancel_with_code(&self, code: CancelCode) -> Result<JsValue, JsError> {
         self.inner
@@ -299,25 +307,25 @@ impl Sas {
             .map_err(Into::into)
     }
 
-    /// Has the SAS verification flow timed out.
+    /// Has the SAS verification flow timed out?
     #[wasm_bindgen(js_name = "timedOut")]
     pub fn timed_out(&self) -> bool {
         self.inner.timed_out()
     }
 
-    /// Are we in a state where we can show the short auth string.
+    /// Are we in a state where we can show the short auth string?
     #[wasm_bindgen(js_name = "canBePresented")]
     pub fn can_be_presented(&self) -> bool {
         self.inner.can_be_presented()
     }
 
-    /// Is the SAS flow done.
+    /// Is the SAS flow done?
     #[wasm_bindgen(js_name = "isDone")]
     pub fn is_done(&self) -> bool {
         self.inner.is_done()
     }
 
-    /// Is the SAS flow canceled.
+    /// Is the SAS flow cancelled?
     #[wasm_bindgen(js_name = "isCancelled")]
     pub fn is_cancelled(&self) -> bool {
         self.inner.is_cancelled()
@@ -326,7 +334,7 @@ impl Sas {
     /// Get the emoji version of the short auth string.
     ///
     /// Returns `undefined` if we can't yet present the short auth string,
-    /// otherwise seven tuples containing the emoji and description.
+    /// otherwise an array of seven `Emoji` objects.
     pub fn emoji(&self) -> Option<Array> {
         Some(
             self.inner
@@ -406,7 +414,7 @@ impl Qr {
         self.inner.has_been_scanned()
     }
 
-    /// Has the scanning of the QR code been confirmed by us.
+    /// Has the scanning of the QR code been confirmed by us?
     #[wasm_bindgen(js_name = "hasBeenConfirmed")]
     pub fn has_been_confirmed(&self) -> bool {
         self.inner.has_been_confirmed()
@@ -431,7 +439,7 @@ impl Qr {
         self.inner.other_device_id().to_owned().into()
     }
 
-    /// Did we initiate the verification request.
+    /// Did we initiate the verification request?
     #[wasm_bindgen(js_name = "weStarted")]
     pub fn we_started(&self) -> bool {
         self.inner.we_started()
@@ -444,26 +452,26 @@ impl Qr {
         self.inner.cancel_info().map(Into::into)
     }
 
-    /// Has the verification flow completed.
+    /// Has the verification flow completed?
     #[wasm_bindgen(js_name = "isDone")]
     pub fn is_done(&self) -> bool {
         self.inner.is_done()
     }
 
-    /// Has the verification flow been cancelled.
+    /// Has the verification flow been cancelled?
     #[wasm_bindgen(js_name = "isCancelled")]
     pub fn is_cancelled(&self) -> bool {
         self.inner.is_cancelled()
     }
 
-    /// Is this a verification that is veryfying one of our own devices.
+    /// Is this a verification that is verifying one of our own devices?
     #[wasm_bindgen(js_name = "isSelfVerification")]
     pub fn is_self_verification(&self) -> bool {
         self.inner.is_self_verification()
     }
 
     /// Have we successfully scanned the QR code and are able to send
-    /// a reciprocation event.
+    /// a reciprocation event?
     pub fn reciprocated(&self) -> bool {
         self.inner.reciprocated()
     }
@@ -490,6 +498,8 @@ impl Qr {
     ///
     /// The `to_bytes` method can be used to instead output the raw
     /// bytes that should be encoded as a QR code.
+    ///
+    /// Returns a `QrCode`.
     #[wasm_bindgen(js_name = "toQrCode")]
     pub fn to_qr_code(&self) -> Result<QrCode, JsError> {
         Ok(self.inner.to_qr_code().map(Into::into)?)
@@ -526,6 +536,9 @@ impl Qr {
     }
 
     /// Confirm that the other side has scanned our QR code.
+    ///
+    /// Returns either an `OutgoingRequest` which should be sent out, or
+    /// `undefined` if the verification is already confirmed.
     #[wasm_bindgen(js_name = "confirmScanning")]
     pub fn confirm_scanning(&self) -> Result<JsValue, JsError> {
         self.inner
@@ -538,6 +551,9 @@ impl Qr {
     }
 
     /// Cancel the verification flow.
+    ///
+    /// Returns either an `OutgoingRequest` which should be sent out, or
+    /// `undefined` if the verification is already cancelled.
     pub fn cancel(&self) -> Result<JsValue, JsError> {
         self.inner
             .cancel()
@@ -551,6 +567,9 @@ impl Qr {
     /// Cancel the verification.
     ///
     /// This cancels the verification with given code.
+    ///
+    /// Returns either an `OutgoingRequest` which should be sent out, or
+    /// `undefined` if the verification is already cancelled.
     #[wasm_bindgen(js_name = "cancelWithCode")]
     pub fn cancel_with_code(&self, code: CancelCode) -> Result<JsValue, JsError> {
         self.inner
@@ -871,19 +890,19 @@ impl VerificationRequest {
         self.inner.cancel_info().map(Into::into)
     }
 
-    /// Has the verification request been answered by another device.
+    /// Has the verification request been answered by another device?
     #[wasm_bindgen(js_name = "isPassive")]
     pub fn is_passive(&self) -> bool {
         self.inner.is_passive()
     }
 
-    /// Is the verification request ready to start a verification flow.
+    /// Is the verification request ready to start a verification flow?
     #[wasm_bindgen(js_name = "isReady")]
     pub fn is_ready(&self) -> bool {
         self.inner.is_ready()
     }
 
-    /// Has the verification flow timed out.
+    /// Has the verification flow timed out?
     #[wasm_bindgen(js_name = "timedOut")]
     pub fn timed_out(&self) -> bool {
         self.inner.timed_out()
@@ -925,30 +944,53 @@ impl VerificationRequest {
             .transpose()
     }
 
-    /// Get the unique ID of this verification request
+    /// Get the unique ID of this verification request.
     #[wasm_bindgen(getter, js_name = "flowId")]
     pub fn flow_id(&self) -> String {
         self.inner.flow_id().as_str().to_owned()
     }
 
-    /// Is this a verification that is veryfying one of our own
-    /// devices.
+    /// Is this a verification that is verifying one of our own
+    /// devices?
     #[wasm_bindgen(js_name = "isSelfVerification")]
     pub fn is_self_verification(&self) -> bool {
         self.inner.is_self_verification()
     }
 
-    /// Did we initiate the verification request.
+    /// Did we initiate the verification request?
     #[wasm_bindgen(js_name = "weStarted")]
     pub fn we_started(&self) -> bool {
         self.inner.we_started()
     }
 
     /// Has the verification flow that was started with this request
-    /// finished.
+    /// finished?
     #[wasm_bindgen(js_name = "isDone")]
     pub fn is_done(&self) -> bool {
         self.inner.is_done()
+    }
+
+    /// Get the current phase of this request.
+    ///
+    /// Returns a `VerificationRequestPhase`.
+    pub fn phase(&self) -> VerificationRequestPhase {
+        self.inner.state().into()
+    }
+
+    /// If this request has transitioned into a concrete verification
+    /// flow (and not yet been completed or cancelled), returns a `Verification`
+    /// object.
+    ///
+    /// Returns: a `Sas`, a `Qr`, or `undefined`.
+    #[wasm_bindgen(js_name = "getVerification")]
+    pub fn get_verification(&self) -> JsValue {
+        let result: Option<JsValue> =
+            if let VerificationRequestState::Transitioned { verification } = self.inner.state() {
+                Verification(verification).try_into().ok()
+            } else {
+                None
+            };
+        result.into()
     }
 
     /// Register a callback which will be called whenever there is an update to
@@ -969,9 +1011,9 @@ impl VerificationRequest {
             stream.for_each(|_| send_change_info_to_callback(callback_ref)).await;
         });
     }
-
+      
     /// Has the verification flow that was started with this request
-    /// been cancelled.
+    /// been cancelled?
     #[wasm_bindgen(js_name = "isCancelled")]
     pub fn is_cancelled(&self) -> bool {
         self.inner.is_cancelled()
@@ -1121,6 +1163,45 @@ impl TryFrom<OutgoingVerificationRequest> for JsValue {
                 JsValue::from(requests::RoomMessageRequest::try_from((request_id, &request))?)
             }
         })
+    }
+}
+
+/// List of VerificationRequestState phases
+#[wasm_bindgen]
+#[derive(Debug, Clone)]
+pub enum VerificationRequestPhase {
+    /// The verification request has been newly created by us.
+    Created = 0,
+
+    /// The verification request was received from the other party.
+    Requested = 1,
+
+    /// The verification request is ready to start a verification flow.
+    Ready = 2,
+
+    /// The verification request has transitioned into a concrete verification
+    /// flow. For example it transitioned into the emoji based SAS
+    /// verification.
+    Transitioned = 3,
+
+    /// The verification flow that was started with this request has finished.
+    Done = 4,
+
+    /// The verification process has been cancelled.
+    Cancelled = 5,
+}
+
+impl From<VerificationRequestState> for VerificationRequestPhase {
+    fn from(value: VerificationRequestState) -> Self {
+        use matrix_sdk_crypto::VerificationRequestState::*;
+        match value {
+            Created { .. } => Self::Created,
+            Requested { .. } => Self::Requested,
+            Transitioned { .. } => Self::Transitioned,
+            Ready { .. } => Self::Ready,
+            Done => Self::Done,
+            Cancelled(_) => Self::Cancelled,
+        }
     }
 }