+++ crates/matrix-sdk-ui/tests/integration/room_list.rs
@@ -239,8 +239,15 @@ async fn test_sync_from_init_to_enjoy() -> Result<(), Error> {
                         [0, 19],
                     ],
                     "required_state": [
+                        ["m.room.avatar", ""],
                         ["m.room.encryption", ""],
+                        ["m.room.power_levels", ""],
                     ],
+                    "filters": {
+                        "is_invite": false,
+                        "is_tombstoned": false,
+                        "not_room_types": ["m.space"],
+                    },
                     "sort": ["by_recency", "by_name"],
                     "timeline_limit": 1,
                 },
@@ -290,6 +297,11 @@ async fn test_sync_from_init_to_enjoy() -> Result<(), Error> {
                     "required_state": [
                         ["m.room.encryption", ""],
                     ],
+                    "filters": {
+                        "is_invite": false,
+                        "is_tombstoned": false,
+                        "not_room_types": ["m.space"],
+                    },
                     "sort": ["by_recency", "by_name"],
                     "timeline_limit": 20,
                 }