+++ crates/matrix-sdk/src/sliding_sync/mod.rs
@@ -1066,7 +1066,7 @@ mod tests {
 
         // If there's a to-device `since` token, we make sure we put the token
         // into the extension config. The rest doesn't need to be re-enabled due to
-        // stickyness.
+        // stickiness.
         let since_token = "since";
         sync.inner.position.write().unwrap().to_device_token = Some(since_token.to_owned());
 
@@ -1185,6 +1185,48 @@ mod tests {
 
         // The sync-loop is actually running.
         assert!(stream.next().await.is_some());
+
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn test_sliding_sync_proxy_url() -> Result<()> {
+        let server = MockServer::start().await;
+        let client = logged_in_client(Some(server.uri())).await;
+
+        {
+            // A server that doesn't expose a sliding sync proxy gets and transmits none, by
+            // default.
+            let sync = client.sliding_sync("no-proxy")?.build().await?;
+
+            assert!(sync.sliding_sync_proxy().is_none());
+        }
+
+        {
+            // The sliding sync builder can be used to customize a proxy, though.
+            let url = Url::parse("https://bar.matrix/").unwrap();
+            let sync =
+                client.sliding_sync("own-proxy")?.sliding_sync_proxy(url.clone()).build().await?;
+            assert_eq!(sync.sliding_sync_proxy(), Some(url));
+        }
+
+        // Set the client's proxy, that will be inherited by sliding sync.
+        let url = Url::parse("https://foo.matrix/").unwrap();
+        client.set_sliding_sync_proxy(Some(url.clone()));
+
+        {
+            // The sliding sync inherits the client's sliding sync proxy URL.
+            let sync = client.sliding_sync("client-proxy")?.build().await?;
+            assert_eq!(sync.sliding_sync_proxy(), Some(url));
+        }
+
+        {
+            // …unless we override it.
+            let url = Url::parse("https://bar.matrix/").unwrap();
+            let sync =
+                client.sliding_sync("own-proxy")?.sliding_sync_proxy(url.clone()).build().await?;
+            assert_eq!(sync.sliding_sync_proxy(), Some(url));
+        }
 
         Ok(())
     }