+++ testing/matrix-sdk-integration-testing/src/helpers.rs
@@ -47,7 +47,7 @@ pub async fn get_client_for_user(username: String, use_sqlite_store: bool) -> Re
     let tmp_dir = tempdir()?;
 
     let client_builder = Client::builder()
-        .user_agent("matrix-sdk-integation-tests")
+        .user_agent("matrix-sdk-integration-tests")
         .homeserver_url(homeserver_url)
         .request_config(RequestConfig::short_retry());
     let client = if use_sqlite_store {