+++ crates/matrix-sdk/CHANGELOG.md
@@ -13,6 +13,8 @@
 - The `Common` methods to retrieve state events can now return a sync or stripped event, so it can be used
   for invited rooms too.
 - Add `Client::subscribe_to_room_updates` and `room::Common::subscribe_to_updates`
+- `Client::rooms` now returns all rooms, even invited, as advertised.
+- Add `Client::rooms_filtered`
 
 # 0.6.2